reviewing finally
-making queries and messages more proper

+++ api/pontusbot.php
@@ -85,8 +85,16 @@ class bot
 			$getStatus = explode("-", $this->test);
 			$count_getstatus = count($getStatus);
 			if ($count_getstatus > 1) {
-				$this->mysqli->query("DELETE  FROM $this->tbl_con WHERE  `user_convo` = '$this->email'");
-				return ["Cleaning all your Conversation...",$this->username.", your conversation has been cleared"];
+				// making checks
+				$making_checks = $this->mysqli->query("SELECT * FROM $this->tbl_con WHERE  `user_convo` = '$this->email'");
+				$count_making_checks = $making_checks->num_rows;
+				if ($count_making_checks > 0 ) {
+					
+					$this->mysqli->query("DELETE  FROM $this->tbl_con WHERE  `user_convo` = '$this->email'");
+					return ["Cleaning all your Conversation...",$this->username.", your conversation has been cleared"];
+				}else{
+					return ["Searching  Conversations...",$this->username.", no conversation was found for this date range"];
+				}
 			}else{
 				return  ["none","Nothing to process, Follow the format given to you when you type help. E.g D-all"];
 			}
@@ -107,8 +115,16 @@ class bot
 				}else{
 					$date = $this->chatDelete(- $status, "-");
 				}
-				$this->mysqli->query("DELETE  FROM $this->tbl_con WHERE `convo_time` = '%$date%' AND  `user_convo` LIKE '$this->email' ");
-				return ["Cleaning  conversation ...",$this->username.", your Conversation Cleared"];
+				// making checks
+				$making_checks = $this->mysqli->query("SELECT * FROM $this->tbl_con WHERE  `user_convo` = '$this->email' AND  `convo_time` LIKE '%$date%'");
+				$count_making_checks = $making_checks->num_rows;
+				if ($count_making_checks > 0 ) {
+					
+					$this->mysqli->query("DELETE  FROM $this->tbl_con WHERE `convo_time` LIKE '%$date%' AND  `user_convo` = '$this->email' ");
+					return ["Cleaning  conversation ...",$this->username.", your Conversation Cleared"];
+				}else{
+					return ["Searching  Conversations...",$this->username.", no conversation was found for this date range"];
+				}
 			}else{
 				return  ["none","Nothing to process, Follow the format given to you when you type help. E.g Delx-five-years-ago"];
 			}
@@ -117,8 +133,15 @@ class bot
 			$count_getstatus = count($getStatus);
 			if ($count_getstatus > 1) {
 			    $date = date("Y-m-d");
-				$this->mysqli->query("DELETE  FROM $this->tbl_con WHERE  `user_convo` = '$this->email' AND `convo_time` LIKE '%$date%' ");
-				return ["Cleaning all your Conversation...",$this->username.", your conversation has been cleared"];
+				// making checks
+				$making_checks = $this->mysqli->query("SELECT * FROM $this->tbl_con WHERE  `user_convo` = '$this->email' AND `convo_time` LIKE '%$date%'");
+				$count_making_checks = $making_checks->num_rows;
+				if ($count_making_checks > 0 ) {
+					$this->mysqli->query("DELETE  FROM $this->tbl_con WHERE  `user_convo` = '$this->email' AND `convo_time` LIKE '%$date%' ");
+					return ["Cleaning todays  Conversation...",$this->username.", your todays conversation has been cleared"];
+				}else{
+					return ["Searching  Conversations...",$this->username.", no conversation was found for this date range"];
+				}
 			}else{
 				return  ["none","Nothing to process, Follow the format given to you when you type help. E.g D-today"];
 			}
